Refactor to be more modular

+++ bootstrap.sh
@@ -4,11 +4,15 @@ setopt EXTENDED_GLOB
 
 FILES=(
     ".bashrc"
+    ".bashrc.machine.sh"
     ".zshrc"
-    ".commonrc"
+    ".zshrc.machine.sh"
+    ".common.sh"
     ".gitconfig"
-    ".lein"
     ".ssh"
+    "leiningen/.lein"
+    "boot/.boot"
+    "tmux/.tmux.conf"
 )
 
 realpath() {
@@ -18,7 +22,9 @@ realpath() {
 for FILE in "${FILES[@]}"
 do
     HOME_FILE="$HOME/`basename $FILE`"
-    echo Creating link $HOME_FILE
-    rm -rf $HOME_FILE 2> /dev/null || true
-    ln -s `realpath $FILE` $HOME_FILE
+    if [[ -e $FILE ]]; then
+        echo Creating link $HOME_FILE
+        rm -f $HOME_FILE 2> /dev/null || true
+        ln -s `realpath $FILE` $HOME_FILE
+    fi
 done
